Refactored errors
