+++ TODO
@@ -17,4 +17,5 @@
 ☐ Refactor error handling to Middleware
 ☐ Refactor use of index.js making it hard to debug or find stuff @high
     ✔ Refactor course domain @done(20-12-05 14:56)
-    ☐ Refactor anything that has code in the index.js
+    ✔ Refactor Database @done(20-12-05 15:19)
+    ✔ Refactor errors @done(20-12-05 15:20)
