+++ src/courses/getCourse.js
@@ -1,6 +1,7 @@
 import courseModel from "../models/course";
 import { removeDatabaseProps } from "../database";
-import { InvalidArgumentError, NotFoundError } from "../errors";
+import { NotFoundError } from "../errors/NotFoundError";
+import { InvalidArgumentError } from "../errors/InvalidArgumentError";
 
 export const getCourse = async (courseId, userId) => {
 	if (!courseId) throw new InvalidArgumentError("courseId is required");
