+++ src/errors/InvalidArgumentError.js
@@ -1,0 +1,1 @@
+export class InvalidArgumentError extends Error { }
