+++ src/errors/NotFoundError.js
@@ -1,0 +1,1 @@
+export class NotFoundError extends Error { }
