+++ src/errors/ValidationError.js
@@ -1,0 +1,6 @@
+export class ValidationError extends Error {
+	constructor(message, errors) {
+		super(message);
+		this.errors = errors;
+	}
+}
